feat: add create scope functionality with machine selection and storage device configuration
