+++ otterscale-web/messages/en.json
@@ -165,5 +165,31 @@
 	"customization_services_description": "Offers custom development, integration services, and professional consulting.",
 	"training_services": "Training Services",
 	"training_services_description": "Offers exclusive training courses for enterprise users.",
+	"create_scope_name": "Name",
+	"create_scope_name_description": "Enter a unique identifier for your scope and prefix for facility naming.",
+	"create_scope_machine": "Machine",
+	"create_scope_machine_description": "Choose a machine for your Ceph and Kubernetes deployment.",
+	"create_scope_machine_select": "Select a machine",
+	"create_scope_storage_devices": "Storage Devices",
+	"create_scope_storage_devices_description": "Configure dedicated storage devices for Ceph.",
+	"create_scope_calico_cidr": "Calico CIDR",
+	"create_scope_calico_cidr_description": "Network range for Kubernetes pod communication.",
+	"create_scope_virtual_ip": "Virtual IP",
+	"create_scope_virtual_ip_description": "High availability IP for Kubernetes control plane.",
+	"architecture": "Architecture",
+	"cpu": "CPU",
+	"memory": "Memory",
+	"system": "System",
+	"vendor": "Vendor",
+	"product": "Product",
+	"version": "Version",
+	"serial": "Serial",
+	"mainboard": "Mainboard",
+	"firmware": "Firmware",
+	"boot_mode": "Boot mode",
+	"date": "Date",
+	"name": "Name",
+	"mac_address": "MAC address",
+	"optional": "Optional",
 	"error_generic": "Something went wrong!"
 }
