+++ otterscale-web/messages/zh-hant.json
@@ -165,5 +165,31 @@
 	"customization_services_description": "提供客製化開發、整合服務和專業顧問諮詢。",
 	"training_services": "教育訓練服務",
 	"training_services_description": "為企業使用者提供專屬培訓課程。",
+	"create_scope_name": "名稱",
+	"create_scope_name_description": "請輸入您範疇的唯一識別碼和設施命名前綴。",
+	"create_scope_machine": "機器",
+	"create_scope_machine_description": "為您的 Ceph 和 Kubernetes 部署選擇一台機器。",
+	"create_scope_machine_select": "選擇機器",
+	"create_scope_storage_devices": "儲存裝置",
+	"create_scope_storage_devices_description": "為 Ceph 設定專用儲存裝置。",
+	"create_scope_calico_cidr": "Calico 無類別域間路由",
+	"create_scope_calico_cidr_description": "Kubernetes Pod 通訊的網路範圍。",
+	"create_scope_virtual_ip": "虛擬 IP",
+	"create_scope_virtual_ip_description": "Kubernetes 控制平面的高可用性 IP。",
+	"architecture": "架構",
+	"cpu": "處理器",
+	"memory": "記憶體",
+	"system": "系統",
+	"vendor": "廠商",
+	"product": "產品",
+	"version": "版本",
+	"serial": "序號",
+	"mainboard": "主機板",
+	"firmware": "韌體",
+	"boot_mode": "開機模式",
+	"date": "日期",
+	"name": "名稱",
+	"mac_address": "MAC 位址",
+	"optional": "選填",
 	"error_generic": "發生錯誤！"
 }
